requirements.txt
pytest
pytest-flask
flask
django
mysql-connector-python
SQLAlchemy
Flask-SQLAlchemy
mock